Added rtu and tcp wrappers.

+++ SharpModbus/ModbusCommand.cs
@@ -5,8 +5,8 @@ namespace SharpModbus
 {
 	public interface ModbusCommand
 	{
-		ushort RequestLength { get; }
-		ushort ResponseLength { get; }
+		int RequestLength { get; }
+		int ResponseLength { get; }
 		void FillRequest(byte[] request, int offset);
 	}
 }
